pyproject.toml
# Copyright 2026 Enactic, Inc.
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openarm_dataset"
version = "0.1.0"
authors = [{name = "Enactic, Inc."}]
license = {text = "Apache-2.0"}
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "numpy",
    "pandas",
    "pillow",
    "pyarrow",
    "pyyaml",
    "rerun-sdk",
    "scipy",
]

[project.urls]
Changelog = "https://github.com/enactic/openarm_dataset/releases"
Issues = "https://github.com/enactic/openarm_dataset/issues"
Repository = "https://github.com/enactic/openarm_dataset.git"

[dependency-groups]
dev = [
    "pytest",
    "ruff == 0.14.11", # sync with .pre-commit-config.yaml
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff.lint]
extend-select = [
  "D",   # pydocstyle
  "UP",
]

[tool.ruff.lint.per-file-ignores]
"!src/**.py" = ["D"]
